requirements.txt
# Core data fetching
fredapi==0.5.2
yfinance==0.2.40

# Data manipulation
pandas==2.2.2
numpy==1.26.4

# AI synthesis
openai==1.35.3

# Web scraping for WGC monitoring
requests==2.32.3
beautifulsoup4==4.12.3

# Plotting (optional, for charts in reports)
matplotlib==3.9.0

# Environment variables
python-dotenv==1.0.1

# Date/time utilities
python-dateutil==2.9.0

